endpoint: provide EndpointAPIManager via hive cell
Currently, some methods on the `Daemon` are used by the Endpoint API
and DeletionQueue.

In preparation for extracting the Endpoint API, this commit extracts
these methods from the `Daemon` and introduces a new Hive component
`EndpointAPIManager` that provides this functionality. It is provided
by the Hive module `endpoint-api`.
