+++ daemon/cmd/daemon.go
@@ -17,7 +17,6 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink"
 
-	"github.com/cilium/cilium/daemon/cmd/cni"
 	agentK8s "github.com/cilium/cilium/daemon/k8s"
 	"github.com/cilium/cilium/pkg/clustermesh"
 	"github.com/cilium/cilium/pkg/controller"
@@ -107,8 +106,9 @@ type Daemon struct {
 	// ipam is the IP address manager of the agent
 	ipam *ipam.IPAM
 
-	endpointCreator endpointcreator.EndpointCreator
-	endpointManager endpointmanager.EndpointManager
+	endpointCreator    endpointcreator.EndpointCreator
+	endpointManager    endpointmanager.EndpointManager
+	endpointAPIManager endpointapi.EndpointAPIManager
 
 	endpointRestoreComplete       chan struct{}
 	endpointInitialPolicyComplete chan struct{}
@@ -128,10 +128,6 @@ type Daemon struct {
 
 	ciliumHealth health.CiliumHealthManager
 
-	// endpointCreations is a map of all currently ongoing endpoint
-	// creation events
-	endpointCreations endpointapi.EndpointCreationManager
-
 	apiLimiterSet *rate.APILimiterSet
 
 	// CIDRs for which identities were restored during bootstrap
@@ -140,9 +136,6 @@ type Daemon struct {
 	// Controllers owned by the daemon
 	controllers *controller.Manager
 	jobGroup    job.Group
-
-	// just used to tie together some status reporting
-	cniConfigManager cni.CNIConfigManager
 
 	// read-only map of all the hive settings
 	settings cellSettings
@@ -320,48 +313,47 @@ func newDaemon(ctx context.Context, cleaner *daemonCleanup, params *daemonParams
 	})
 
 	d := Daemon{
-		ctx:               ctx,
-		logger:            params.Logger,
-		clientset:         params.Clientset,
-		db:                params.DB,
-		mtuConfig:         params.MTU,
-		directRoutingDev:  params.DirectRoutingDevice,
-		nodeAddressing:    params.NodeAddressing,
-		routes:            params.Routes,
-		devices:           params.Devices,
-		nodeAddrs:         params.NodeAddrs,
-		nodeDiscovery:     params.NodeDiscovery,
-		nodeLocalStore:    params.LocalNodeStore,
-		endpointCreations: params.EndpointCreations,
-		apiLimiterSet:     params.APILimiterSet,
-		controllers:       controller.NewManager(),
-		jobGroup:          params.JobGroup,
+		ctx:              ctx,
+		logger:           params.Logger,
+		clientset:        params.Clientset,
+		db:               params.DB,
+		mtuConfig:        params.MTU,
+		directRoutingDev: params.DirectRoutingDevice,
+		nodeAddressing:   params.NodeAddressing,
+		routes:           params.Routes,
+		devices:          params.Devices,
+		nodeAddrs:        params.NodeAddrs,
+		nodeDiscovery:    params.NodeDiscovery,
+		nodeLocalStore:   params.LocalNodeStore,
+		apiLimiterSet:    params.APILimiterSet,
+		controllers:      controller.NewManager(),
+		jobGroup:         params.JobGroup,
 		// **NOTE** The global identity allocator is not yet initialized here; that
 		// happens below via InitIdentityAllocator(). Only the local identity
 		// allocator is initialized here.
-		identityAllocator: params.IdentityAllocator,
-		ipcache:           params.IPCache,
-		policy:            params.Policy,
-		idmgr:             params.IdentityManager,
-		cniConfigManager:  params.CNIConfigManager,
-		clustermesh:       params.ClusterMesh,
-		monitorAgent:      params.MonitorAgent,
-		svc:               params.ServiceManager,
-		settings:          params.Settings,
-		bwManager:         params.BandwidthManager,
-		endpointCreator:   params.EndpointCreator,
-		endpointManager:   params.EndpointManager,
-		endpointMetadata:  params.EndpointMetadata,
-		k8sWatcher:        params.K8sWatcher,
-		k8sSvcCache:       params.K8sSvcCache,
-		ipam:              params.IPAM,
-		wireguardAgent:    params.WGAgent,
-		orchestrator:      params.Orchestrator,
-		lrpManager:        params.LRPManager,
-		maglevConfig:      params.MaglevConfig,
-		lbConfig:          params.LBConfig,
-		ciliumHealth:      params.CiliumHealth,
-		statusCollector:   params.StatusCollector,
+		identityAllocator:  params.IdentityAllocator,
+		ipcache:            params.IPCache,
+		policy:             params.Policy,
+		idmgr:              params.IdentityManager,
+		clustermesh:        params.ClusterMesh,
+		monitorAgent:       params.MonitorAgent,
+		svc:                params.ServiceManager,
+		settings:           params.Settings,
+		bwManager:          params.BandwidthManager,
+		endpointCreator:    params.EndpointCreator,
+		endpointManager:    params.EndpointManager,
+		endpointMetadata:   params.EndpointMetadata,
+		endpointAPIManager: params.EndpointAPIManager,
+		k8sWatcher:         params.K8sWatcher,
+		k8sSvcCache:        params.K8sSvcCache,
+		ipam:               params.IPAM,
+		wireguardAgent:     params.WGAgent,
+		orchestrator:       params.Orchestrator,
+		lrpManager:         params.LRPManager,
+		maglevConfig:       params.MaglevConfig,
+		lbConfig:           params.LBConfig,
+		ciliumHealth:       params.CiliumHealth,
+		statusCollector:    params.StatusCollector,
 	}
 
 	// initialize endpointRestoreComplete channel as soon as possible so that subsystems
